Refactoring

+++ createmusic/views.py
@@ -119,6 +119,9 @@ class MusicMergeView(View):
     def post(self, request, *args, **kwargs):
         submusic_pk = request.POST.get('data')
         music_pk = self.kwargs.get(self.pk_url_kwarg)
+        if Music.objects.get(id=music_pk).owner == self.request.user:
+            message = '권한이 없습니다.'
+            return JsonResponse(status=403, data={'message': message})
 
         submusic = get_object_or_404(SubMusic, pk=submusic_pk)
         submusic.status = 0
@@ -127,7 +130,7 @@ class MusicMergeView(View):
         return JsonResponse(status=200, data={'message': message})
 
     def get(self, request, *args, **kwrags):
-        message = '잘못 된 접근입니다'
+        message = '잘못된 접근입니다'
         return JsonResponse(status=405, data={'message': message})
 
 
@@ -139,6 +142,10 @@ class SubMusicDeleteView(View):
         submusic_pk = request.POST.get('data')
         music_pk = self.kwargs.get(self.pk_url_kwargs)
 
+        if Music.objects.get(id=music_pk).owner == self.request.user:
+            message = '권한이 없습니다.'
+            return JsonResponse(status=403, data={'message': message})
+
         sub_music = get_object_or_404(SubMusic, pk=submusic_pk)
         sub_music.delete()
 
@@ -146,7 +153,7 @@ class SubMusicDeleteView(View):
         return JsonResponse(status=200, data={'message': message})
 
     def get(self, request, *args, **kwargs):
-        message = '잘못 된 접근입니다.'
+        message = '잘못된 접근입니다.'
         return JsonResponse(status=405, data={'message': message})
